chore: refactor map to separate zoom and options
