+++ src/client/components/galaxymap/MapOptions.tsx
@@ -1,0 +1,31 @@
+import { useState } from "react";
+import Expandable from "../oodreact/Expandable";
+import styles from "./map.module.css";
+
+export interface IMapOptions {
+    hidePlanetLabels?: boolean;
+    hideSpacelaneLabels?: boolean;
+    showAllPlanets?: boolean;
+    showAllSpacelanes?: boolean;
+}
+
+interface IMapOption<T> {
+    currentValue: T;
+    setValue: (value: T) => void;
+    label: string;
+    inputType: string;
+}
+
+type MapOption = IMapOption<boolean>
+
+interface IMapOptionProps {
+    mapOptions: MapOption[]
+}
+
+export function MapOptions(props: IMapOptionProps) {
+    return (
+    <Expandable className={styles.optionsWindow} title="Map Options">
+        {props.mapOptions.map((p, i) => <span key={i}><input type={p.inputType} onChange={(event) => p.setValue(event.target.checked)} /><label>{p.label}</label></span>)}
+    </Expandable>
+    );
+}
